requirements.txt
# Promptbench will not compilles above version 11.4

aiohttp
ollama
python-dotenv
hikari
numpy
scikit-learn
sentence-transformers
transformers 
torch
promptbench
pytextrank
spacy
pytextrank
PyYAML
tqdm
pandas